# 8302877: Speed up latin1 case conversions