--- a/test/micro/org/openjdk/bench/java/lang/Characters.java
+++ b/test/micro/org/openjdk/bench/java/lang/Characters.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -27,11 +27,13 @@ package org.openjdk.bench.java.lang;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
+import org.openjdk.jmh.annotations.Level;
 import org.openjdk.jmh.annotations.Measurement;
 import org.openjdk.jmh.annotations.Mode;
 import org.openjdk.jmh.annotations.OutputTimeUnit;
 import org.openjdk.jmh.annotations.Param;
 import org.openjdk.jmh.annotations.Scope;
+import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.annotations.Warmup;
 
@@ -79,5 +81,48 @@ public class Characters {
             Character.codePointOf("Latin Capital Letter B with hook");
         }
 
+    }
+
+    @BenchmarkMode(Mode.AverageTime)
+    @OutputTimeUnit(TimeUnit.NANOSECONDS)
+    @State(Scope.Thread)
+    @Warmup(iterations = 5, time = 1)
+    @Measurement(iterations = 5, time = 1)
+    @Fork(3)
+    public static class Latin1CaseFolding {
+        @Param({
+                "low",   // 0x09 pre A
+                "A",  // 0x41 uppercase A
+                "a",  // 0x61 lowercase a
+                "A-grave", // 0xC0 uppercase A-grave
+                "a-grave", // 0xE0 lowercase a-grave
+                "mu", // 0xB5 lowercase 'my'
+                "yD" // 0xFF lowercase 'y with Diaeresis'
+        })
+        private String codePoint;
+        private int cp;
+
+        @Setup(Level.Trial)
+        public void setup() {
+            cp = switch (codePoint) {
+                case "low" -> 0x09;
+                case "A" -> 0x41;
+                case "a" -> 0x61;
+                case "A-grave" -> 0xC0;
+                case "a-grave" -> 0xE0;
+                case "yD" -> 0xE0;
+                case "mu" -> 0xFF;
+                default -> throw new IllegalArgumentException("Unknown character: " + codePoint);
+            };
+        }
+        @Benchmark
+        public int toUpperCase() {
+            return Character.toUpperCase(cp);
+        }
+
+        @Benchmark
+        public int toLowerCase() {
+            return Character.toLowerCase(cp);
+        }
     }
 }