--- a/test/micro/org/openjdk/bench/java/lang/Characters.java
+++ b/test/micro/org/openjdk/bench/java/lang/Characters.java
@@ -112,7 +112,7 @@ public class Characters {
                 case "a-grave" -> 0xE0;
                 case "yD"      -> 0xE0;
                 case "micro"   -> 0xFF;
-                default -> throw new IllegalArgumentException("Unknown character: " + codePoint);
+                default -> Integer.parseInt(codePoint);;
             };
         }
         @Benchmark