@eirbjo 2023-02-17
Benchmark results:

Baseline:

```
Benchmark                                 (codePoint)  Mode  Cnt  Score   Error  Units
Characters.Latin1CaseFolding.toLowerCase          low  avgt   15  1.267 ± 0.013  ns/op
Characters.Latin1CaseFolding.toLowerCase            A  avgt   15  1.657 ± 0.011  ns/op
Characters.Latin1CaseFolding.toLowerCase            a  avgt   15  1.258 ± 0.005  ns/op
Characters.Latin1CaseFolding.toLowerCase      A-grave  avgt   15  1.656 ± 0.011  ns/op
Characters.Latin1CaseFolding.toLowerCase      a-grave  avgt   15  1.270 ± 0.023  ns/op
Characters.Latin1CaseFolding.toLowerCase           mu  avgt   15  1.261 ± 0.006  ns/op
Characters.Latin1CaseFolding.toLowerCase           yD  avgt   15  1.260 ± 0.005  ns/op
Characters.Latin1CaseFolding.toUpperCase          low  avgt   15  1.284 ± 0.043  ns/op
Characters.Latin1CaseFolding.toUpperCase            A  avgt   15  1.264 ± 0.008  ns/op
Characters.Latin1CaseFolding.toUpperCase            a  avgt   15  1.818 ± 0.016  ns/op
Characters.Latin1CaseFolding.toUpperCase      A-grave  avgt   15  1.261 ± 0.015  ns/op
Characters.Latin1CaseFolding.toUpperCase      a-grave  avgt   15  1.822 ± 0.013  ns/op
Characters.Latin1CaseFolding.toUpperCase           mu  avgt   15  1.823 ± 0.006  ns/op
Characters.Latin1CaseFolding.toUpperCase           yD  avgt   15  1.822 ± 0.008  ns/op
```

PR:

```
Benchmark                                 (codePoint)  Mode  Cnt  Score   Error  Units
Characters.Latin1CaseFolding.toLowerCase          low  avgt   15  0.878 ± 0.005  ns/op
Characters.Latin1CaseFolding.toLowerCase            A  avgt   15  1.038 ± 0.009  ns/op
Characters.Latin1CaseFolding.toLowerCase            a  avgt   15  1.036 ± 0.007  ns/op
Characters.Latin1CaseFolding.toLowerCase      A-grave  avgt   15  1.357 ± 0.015  ns/op
Characters.Latin1CaseFolding.toLowerCase      a-grave  avgt   15  1.352 ± 0.003  ns/op
Characters.Latin1CaseFolding.toLowerCase           mu  avgt   15  1.273 ± 0.002  ns/op
Characters.Latin1CaseFolding.toLowerCase           yD  avgt   15  1.352 ± 0.004  ns/op
Characters.Latin1CaseFolding.toUpperCase          low  avgt   15  0.880 ± 0.013  ns/op
Characters.Latin1CaseFolding.toUpperCase            A  avgt   15  0.920 ± 0.071  ns/op
Characters.Latin1CaseFolding.toUpperCase            a  avgt   15  1.055 ± 0.013  ns/op
Characters.Latin1CaseFolding.toUpperCase      A-grave  avgt   15  1.394 ± 0.010  ns/op
Characters.Latin1CaseFolding.toUpperCase      a-grave  avgt   15  1.391 ± 0.009  ns/op
Characters.Latin1CaseFolding.toUpperCase           mu  avgt   15  1.597 ± 0.021  ns/op
Characters.Latin1CaseFolding.toUpperCase           yD  avgt   15  1.354 ± 0.003  ns/op
```