@eirbjo 2023-02-18
> Looks good to me.

Thanks for looking! Would you also be willing to file a JBS issue for this?