@eirbjo 2023-02-21
A site note: Early and crude experiements using the Vector API indicate that the 'oldest ASCII trick in the book' vectorizes pretty well.

Here's a benchmark comparing the strings "helloworld" and "HelloWorld" 1024 times, followed by either 'A' or 'B' (to force a an expensive mismatch):

```
Benchmark                    (size)  Mode  Cnt     Score    Error  Units
EqualsIgnoreCase.scalar        1024  avgt   15  6225.624 ± 89.182  ns/op
EqualsIgnoreCase.vectorized    1024  avgt   15  1246.110 ± 14.767  ns/op
```

I have the feeling that most case-insensitive comparisons are pretty short, so not sure how useful this is IRL.