@eirbjo 2023-02-21
> The testcase change is failing to compile:
> 
> ```
> open/test/micro/org/openjdk/bench/java/lang/Characters.java:115: error: case, default, or '}' expected
> ```

Darn, this was rather embarrassing!

Since this was just integrated, I guess I'll need to open a new PR for the fix?