@eirbjo 2023-02-21
I created https://github.com/openjdk/jdk/pull/12701 for this.

Could someone please file a JBS?