@naotoj 2023-02-21
https://bugs.openjdk.org/browse/JDK-8303033